scripts/requirements.txt
numpy==2.2.5
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
